requirements.txt
attrs==25.3.0
blinker==1.8.2
branca==0.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
fiona==1.10.1
flask==3.0.3
Flask-Cors==5.0.0
folium==0.18.0
fonttools==4.57.0
geopandas==0.13.2
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mysql-connector==2.2.9
mysql-connector-python==9.0.0
numpy==1.24.4
packaging==25.0
pandas==2.0.3
pillow==10.4.0
pyparsing==3.1.4
pyproj==3.5.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
shapely==2.0.7
six==1.17.0
tzdata==2025.2
urllib3==2.2.3
werkzeug==3.0.6
xyzservices==2025.4.0
zipp==3.20.2
